requirements.txt
insightface>=0.7.3
opencv-python>=4.8.0
numpy>=1.26.0
Pillow>=10.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
pyinstaller>=6.0.0
pillow-heif>=0.18.0
